requirements.txt
# Core dependencies
pandas>=2.0.0
aiohttp>=3.9.0

# FastAPI and related
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Environment and utilities
python-dotenv>=1.0.0
requests>=2.31.0